Split utility containers into seperate file
This, along with moving objects from snakebids.types into
snakebids.utils.containers, should clean up the code and solve some
of the circular import problems
