+++ snakebids/tests/test_utils.py
@@ -12,13 +12,8 @@ from hypothesis import assume, given
 from hypothesis import strategies as st
 
 import snakebids.tests.strategies as sb_st
-from snakebids.utils.utils import (
-    ImmutableList,
-    MultiSelectDict,
-    RegexContainer,
-    get_wildcard_dict,
-    matches_any,
-)
+from snakebids.utils.containers import ImmutableList, MultiSelectDict, RegexContainer
+from snakebids.utils.utils import get_wildcard_dict, matches_any
 
 
 @st.composite
